se eliminaron los .css de la seccion de pages. se comenzo a mejorar el responsive. se mejoro el responsive del Hero
